+++ R/univariate.R
@@ -134,6 +134,15 @@
 #'   \code{DataFrame}, otherwise a list each element of which is the results for
 #'   each feature. For \code{run*}, a \code{SpatialFeatureExperiment} object
 #'   with the results added. See Details for where the results are stored.
+#' @references
+#' Cliff, A. D., Ord, J. K. 1981 Spatial processes, Pion, p. 17.
+#'
+#' Anselin, L. (1995), Local Indicators of Spatial Association—LISA. Geographical Analysis, 27: 93-115. doi:10.1111/j.1538-4632.1995.tb00338.x
+#'
+#' Ord, J. K., & Getis, A. 2012. Local spatial heteroscedasticity (LOSH), The Annals of Regional Science, 48 (2), 529–539.
+#'
+#' Ord, J. K. and Getis, A. 1995 Local spatial autocorrelation statistics: distributional issues and an application. Geographical Analysis, 27, 286–306
+#'
 #' @name calculateUnivariate
 #' @aliases calculateMoransI
 #' @importFrom BiocParallel SerialParam bplapply
@@ -208,7 +217,9 @@ setMethod(
              BPPARAM = SerialParam(),
              zero.policy = NULL, returnDF = TRUE, p.adjust.method = "BH",
              name = NULL, ...) {
-        if (info(type, "package") %in% c("spdep", "Voyager", NA)) {
+        if (info(type, "variate") != "uni")
+            stop("`type` must be a univariate method.")
+        if (!info(type, "package") %in% c("spdep", "Voyager", NA)) {
             rlang::check_installed(info(type, "package"))
         }
         other_args <- list(...)