+++ tests/testthat/test-multivariate.R
@@ -7,6 +7,12 @@ sfe <- logNormCounts(sfe)
 inds <- order(Matrix::rowSums(logcounts(sfe)), decreasing = TRUE)[1:50]
 mat <- logcounts(sfe)[inds,]
 g <- colGraph(sfe, "visium") <- findVisiumGraph(sfe)
+
+test_that("Error when a univariate method is used", {
+    expect_error(calculateMultivariate(mat, type = "moran", listw = g),
+                 "`type` must be a multivariate method.")
+})
+
 test_that("Correct output structure of multispati_rsp", {
     out <- multispati_rsp(t(mat), listw = g, nfposi = 10, nfnega = 10)
     expect_equal(colnames(out), paste0("PC", 1:20))
@@ -91,6 +97,51 @@ test_that("Correctly add results to the SFE object", {
                            nfnega = 10)
     out <- reducedDim(sfe, "multispati")
     expect_equal(out, ref)
+    expect_message({
+        sfe <- runMultivariate(sfe, "multispati", subset_row = inds, nfposi = 10,
+                               nfnega = 10, dest = "colData")
+    }, "Matrix or array outputs can only be stored in reducedDims.")
 })
-# 6. Correct output structure when output is a vector
-# including the message on reducedDim
+
+test_that("When output is a vector", {
+    out <- calculateMultivariate(sfe, "localC_multi", colGraphName = "visium",
+                                 subset_row = inds)
+    expect_vector(out)
+    sfe <- runMultivariate(sfe, "localC_multi", colGraphName = "visium",
+                           subset_row = inds, dest = "colData")
+    expect_true("localC_multi" %in% names(colData(sfe)))
+    expect_equal(sfe$localC_multi, out)
+    expect_message({
+        sfe <- runMultivariate(sfe, "localC_multi", colGraphName = "visium",
+                               subset_row = inds, dest = "reducedDim")
+    }, "Vector output can only be stored in colData.")
+})
+
+names_expect_lc <- c(
+    "localC_perm_multi", "E.Ci", "Var.Ci", "Z.Ci", "Pr(z != E(Ci))",
+    "Pr(z != E(Ci)) Sim", "Pr(folded) Sim", "Skewness",
+    "Kurtosis", "-log10p Sim", "-log10p_adj Sim", "cluster"
+)
+test_that("When output is a data frame", {
+    df <- calculateMultivariate(sfe, "localC_perm_multi", colGraphName = "visium",
+                                subset_row = inds, nsim = 99)
+    expect_s3_class(df, "data.frame")
+    expect_equal(names(df), names_expect_lc)
+    expect_equal(nrow(df), ncol(sfe))
+})
+
+test_that("Add data frame output to colData", {
+    sfe <- runMultivariate(sfe, "localC_perm_multi", colGraphName = "visium",
+                           subset_row = inds, nsim = 99, dest = "colData")
+    names_expect <- names_expect_lc
+    names_expect[-1] <- paste("localC_perm_multi", names_expect[-1], sep = "_")
+    expect_true(all(names_expect %in% names(colData(sfe))))
+})
+
+test_that("Add data frame output to reducedDim", {
+    sfe <- runMultivariate(sfe, "localC_perm_multi", colGraphName = "visium",
+                           subset_row = inds, nsim = 99, dest = "reducedDim")
+    df <- reducedDim(sfe, "localC_perm_multi")
+    expect_s3_class(df, "data.frame")
+    expect_equal(names(df), names_expect_lc)
+})