+++ tests/testthat/test-univariate.R
@@ -9,6 +9,13 @@ mat <- assay(sfe, "counts")
 mat1 <- mat[, colData(sfe)$sample_id == "sample01"]
 
 out_m <- calculateMoransI(mat1, listw = colGraph(sfe, "visium", sample_id = "sample01"))
+test_that("Error when a multivariate method is used", {
+    expect_error({
+        calculateUnivariate(mat1, type = "multispati",
+                            listw = colGraph(sfe, "visium", sample_id = "sample01"))
+    }, "`type` must be a univariate method.")
+})
+
 test_that("Correct structure of calculateMoransI output (matrix)", {
     expect_s4_class(out_m, "DataFrame")
     expect_equal(names(out_m), c("moran", "K"))
@@ -244,7 +251,7 @@ test_that("DataFrame output for localmoran", {
 names_expect_lg <- c(
     "localG", "Gi", "E.Gi", "Var.Gi", "StdDev.Gi", "Pr(z != E(Gi))",
     "Pr(z != E(Gi)) Sim", "Pr(folded) Sim", "Skewness",
-    "Kurtosis", "-log10p Sim", "-log10p_adj Sim"
+    "Kurtosis", "-log10p Sim", "-log10p_adj Sim", "cluster"
 )
 test_that("DataFrame output for localG_perm", {
     out <- calculateUnivariate(mat1,
@@ -252,7 +259,7 @@ test_that("DataFrame output for localG_perm", {
         type = "localG_perm"
     )
     expect_s4_class(out, "DFrame")
-    expect_true(all(vapply(out, is.matrix, FUN.VALUE = logical(1))))
+    expect_true(all(vapply(out, is.data.frame, FUN.VALUE = logical(1))))
     expect_equal(colnames(out[[1]]), names_expect_lg)
     expect_equal(names(out), rownames(mat1))
     expect_equal(nrow(out), ncol(mat1))